requirements.txt
asgiref==3.8.1
contourpy==1.1.1
cycler==0.12.1
Django==4.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
fonttools==4.53.1
importlib_resources==6.4.4
inflection==0.5.1
kiwisolver==1.4.7
matplotlib==3.7.5
mysqlclient==2.0.3
numpy==1.24.4
packaging==24.1
Pillow==10.4.0
PyJWT==2.8.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
zipp==3.20.1
